Refactor wpt harness
