+++ tests/wpt/FileAPI.blob.test.ts
@@ -1,4 +1,3 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./FileAPI.harness.js";
+import { runSuite, runTestDynamic } from "./FileAPI.harness.js";
 
 runSuite(import.meta.url, runTestDynamic);
