+++ tests/wpt/WebCryptoAPI.generateKey.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-util.js";
-import { runTestDynamic } from "./WebCryptoAPI.harness.js";
+import { runSuite, runTestDynamic } from "./WebCryptoAPI.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "successes_RSA-OAEP.https.any.js", // Error: Test timed out after 5000ms
