+++ tests/wpt/WebCryptoAPI.serialization.test.ts
@@ -1,4 +1,3 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./WebCryptoAPI.harness.js";
+import { runSuite, runTestDynamic } from "./WebCryptoAPI.harness.js";
 
 runSuite(import.meta.url, runTestDynamic);
