+++ tests/wpt/WebCryptoAPI.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./WebCryptoAPI.harness.js";
+import { runSuite, runTestDynamic } from "./WebCryptoAPI.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "idlharness.https.any.js", // ReferenceError: idl_test is not defined
