+++ tests/wpt/console.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./console.harness.js";
+import { runSuite, runTestDynamic } from "./console.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "console-log-symbol.any.js", // Error: Test timed out after 5000ms
