+++ tests/wpt/fetch.api.body.test.ts
@@ -1,4 +1,3 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./fetch.harness.js";
+import { runSuite, runTestDynamic } from "./fetch.harness.js";
 
 runSuite(import.meta.url, runTestDynamic);
