+++ tests/wpt/fetch.api.headers.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./fetch.harness.js";
+import { runSuite, runTestDynamic } from "./fetch.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "header-values-normalize.any.js", // TypeError: cannot read property 'isWorker' of undefined
