+++ tests/wpt/hr-time.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./hr-time.harness.js";
+import { runSuite, runTestDynamic } from "./hr-time.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "idlharness.any.js", // ReferenceError: idl_test is not defined
