+++ tests/wpt/streams.piping.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./streams.harness.js";
+import { runSuite, runTestDynamic } from "./streams.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "general-addition.any.js", // waiting on resolution of https://github.com/whatwg/streams/issues/1243.
