+++ tests/wpt/streams.transferable.test.ts
@@ -1,4 +1,3 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./streams.harness.js";
+import { runSuite, runTestDynamic } from "./streams.harness.js";
 
 runSuite(import.meta.url, runTestDynamic);
