+++ tests/wpt/test262/intl402/Temporal.ZonedDateTime.test.ts
@@ -1,4 +1,3 @@
-import { runSuite } from "../_harness-test262.js";
-import { runTestDynamic } from "./intl402.harness.js";
+import { runSuite, runTestDynamic } from "./intl402.harness.js";
 
 runSuite(import.meta.url, runTestDynamic);
