+++ tests/wpt/url.test.ts
@@ -1,5 +1,4 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./url.harness.js";
+import { runSuite, runTestDynamic } from "./url.harness.js";
 
 runSuite(import.meta.url, runTestDynamic, [
   "historical.any.js", // TypeError: cannot read property 'isWindow' of undefined
