+++ tests/wpt/webidl-harness.js
@@ -1,6 +1,9 @@
-import { makeRunner } from "./_harness-wpt.js";
+import {
+  makeRunnerWPT as makeRunner,
+  runSuiteWPT as runSuite,
+} from "./_harness-util.js";
 
-export const runTestDynamic = makeRunner({
+const runTestDynamic = makeRunner({
   context: () => ({
     scripts: [
       "encoding/resources/encodings.js",
@@ -8,3 +11,5 @@ export const runTestDynamic = makeRunner({
     ],
   }),
 });
+
+export { runSuite, runTestDynamic };
