+++ tests/wpt/webidl.ecmascript-binding.es-exceptions.test.ts
@@ -1,4 +1,3 @@
-import { runSuite } from "./_harness-wpt.js";
-import { runTestDynamic } from "./webidl-harness.js";
+import { runSuite, runTestDynamic } from "./webidl-harness.js";
 
 runSuite(import.meta.url, runTestDynamic);
